backend/requirements.txt
Flask
Pillow
pytz
gunicorn
